// src/types/types.ts
export interface Teacher {
  id: number
  name: string
  subject: string
  category: string
  rating: number
  reviews: number
  price: string
  location: string
  experience: string
  image: string
  available: string
  certified: boolean
  bio: string
  students: number
  discount?: string
  schedule: string
}